refactor: Update code for casting and formatting ♻️
Casting for the context variable in the subghz application was updated to include explicit type casting. Removed unnecessary state assignment in the subghz_scene_rpc.c file. Cleaned up frequency formatting. Furthermore, unnecessary imports were cleaned up and some constant values were updated for accuracy.

+++ applications/settings/notification_settings/notification_settings_app.c
@@ -109,7 +109,7 @@ static void contrast_changed(VariableItem* item) {
     uint8_t index = variable_item_get_current_value_index(item);
 
     variable_item_set_current_value_text(item, contrast_text[index]);
-    app->notification->settings.contrast = contrast_value[index];
+    app->notification->settings.contrast = (int8_t)contrast_value[index];
     notification_message(app->notification, &sequence_lcd_contrast_update);
 }
