+++ applications/system/js_app/js_thread.c
@@ -198,7 +198,7 @@ static void js_require(struct mjs* mjs) {
 static void js_global_to_string(struct mjs* mjs) {
     double num = mjs_get_double(mjs, mjs_arg(mjs, 0));
     char tmp_str[] = "-2147483648";
-    itoa(num, tmp_str, 10);
+    itoa((int32_t)num, tmp_str, 10);
     mjs_val_t ret = mjs_mk_string(mjs, tmp_str, ~0, true);
     mjs_return(mjs, ret);
 }
@@ -206,7 +206,7 @@ static void js_global_to_string(struct mjs* mjs) {
 static void js_global_to_hex_string(struct mjs* mjs) {
     double num = mjs_get_int(mjs, mjs_arg(mjs, 0));
     char tmp_str[] = "-FFFFFFFF";
-    itoa(num, tmp_str, 16);
+    itoa((int32_t)num, tmp_str, 16);
     mjs_val_t ret = mjs_mk_string(mjs, tmp_str, ~0, true);
     mjs_return(mjs, ret);
 }
